requirements.txt
mypy==1.8.0
django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.1
pylint-django
termcolor==2.4.0            